+++ metadata/__init__.py
@@ -1,0 +1,3 @@
+from .queue import enqueue_metadata
+
+__all__ = ["enqueue_metadata"]